Ensure protected functions are not exposed in history component
